Add reflect type assertion
