+++ js/modules/k6/grpc/client.go
@@ -58,6 +58,7 @@ import (
 	"go.k6.io/k6/lib/metrics"
 	"go.k6.io/k6/lib/types"
 	"go.k6.io/k6/stats"
+	reflectpb "google.golang.org/grpc/reflection/grpc_reflection_v1alpha"
 )
 
 //nolint: lll
@@ -230,7 +231,12 @@ func (c *Client) Connect(ctxPtr *context.Context, addr string, params map[string
 				return false, fmt.Errorf("invalid timeout value: %w", err)
 			}
 		case "reflect":
-			reflect, _ = v.(bool)
+			var ok bool
+			reflect, ok = v.(bool)
+			if !ok {
+				return false, fmt.Errorf("invalid reflect value %w", v)
+			}
+
 		default:
 			return false, fmt.Errorf("unknown connect param: %q", k)
 		}
@@ -303,6 +309,53 @@ func (c *Client) Connect(ctxPtr *context.Context, addr string, params map[string
 		}
 	}
 	return true, nil
+}
+
+// reflect will use the grpc reflection api to make the file descriptors available to request.
+// It is called in the connect function the first time the Client.Connect function is called.
+func (c *Client) reflect(ctxPtr *context.Context) error {
+	client := reflectpb.NewServerReflectionClient(c.conn)
+	methodClient, err := client.ServerReflectionInfo(*ctxPtr)
+	if err != nil {
+		return err
+	}
+	req := &reflectpb.ServerReflectionRequest{MessageRequest: &reflectpb.ServerReflectionRequest_ListServices{}}
+	if err = methodClient.Send(req); err != nil {
+		return err
+	}
+	resp, err := methodClient.Recv()
+	if err != nil {
+		return err
+	}
+	listResp := resp.GetListServicesResponse()
+	if listResp == nil {
+		return fmt.Errorf("can't list services")
+	}
+	fdset := &descriptorpb.FileDescriptorSet{}
+	for _, service := range listResp.GetService() {
+		req = &reflectpb.ServerReflectionRequest{
+			MessageRequest: &reflectpb.ServerReflectionRequest_FileContainingSymbol{
+				FileContainingSymbol: service.GetName(),
+			},
+		}
+		if err = methodClient.Send(req); err != nil {
+			return err
+		}
+		resp, err = methodClient.Recv()
+		if err != nil {
+			return fmt.Errorf("error listing methods on '%s': %w", service, err)
+		}
+		fdResp := resp.GetFileDescriptorResponse()
+		for _, f := range fdResp.GetFileDescriptorProto() {
+			a := &descriptorpb.FileDescriptorProto{}
+			if err = proto.Unmarshal(f, a); err != nil {
+				return err
+			}
+			fdset.File = append(fdset.File, a)
+		}
+	}
+	_, err = c.convertToMethodInfo(fdset)
+	return err
 }
 
 // Invoke creates and calls a unary RPC by fully qualified method name
